Refactor temp key warning
This PR extracts the temp key warning as a separate component.
This warning will subsequently be used on additional screens.

The next location this warning will be displayed is the finish registration
screen that displays the identity number after successful registration.
